backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
google-genai==0.3.0
pinecone-client==5.0.1
pypdf==4.2.0
python-docx==1.1.2
pillow==10.3.0
httpx==0.27.0
aiofiles==23.2.1
